Add files via upload
The data consists of 1000 individual claims. The most important variable of interest is fraud_reported. This variable is labeled 1 if a certain claim is
reported to be fraudulent and 0 otherwise. Each claim in the data is described by 40 different attributes, which are represented as columns. We can divide
the descriptive attributes into four main categories: the insured person, the policy of that person, the description of the incident and characteristics of the car
which is involved in the incident.